+++ tests/ut/plugins/UtSourceTest1.h
@@ -33,7 +33,7 @@ public:
 
     Status Read(std::shared_ptr<Buffer> &buffer, size_t expectedLen) override;
 
-    Status GetSize(size_t &size) override;
+    Status GetSize(uint64_t &size) override;
 
     Status IsSeekable(bool& seekable);
